requirements.txt
pymongo==4.8.0
python_dotenv==1.0.1
fastapi==0.111.1